requirements.txt

fastapi
uvicorn
python-dotenv
openai